chore: keep the follow-up under the lint gates
- split indexing.rs (prototype-index latches, strict-store throwers, keys-array
  slot helpers -> array/indexing_support.rs), array/tests.rs (store-lane tests
  -> array/strict_store_tests.rs) and gc/barrier/mod.rs (slot-form barrier
  entry points -> gc/barrier_store.rs, where their runtime twin already lives)
  back under the 2000-line cap; pure moves plus visibility.
- push_pop::resolved_plain_array_flags goes through array_gc_header instead of
  a hand-rolled address floor (addr-class ratchet).
- cargo fmt over the files the branch touched.
