+++ crates/perry-runtime/src/array/header.rs
@@ -493,13 +493,6 @@ pub(crate) unsafe fn array_named_property_get_by_name(
     })
 }
 
-/// Whether this Array owns any side-table properties.
-///
-/// Numeric properties normally live in dense element storage, but a far
-/// sparse index can enter this table and later fall below a grown capacity.
-/// Bulk element operations use this predicate to decline a dense-only path
-/// instead of leaving that second representation observable.
-#[inline]
 /// Does this (already resolved) array head carry named properties in the
 /// side table? Answered by the monotone latch first: until some array has
 /// taken a named property, the table has always been empty.
